+++ src/frontend/apps/drive/src/pages/index.tsx
@@ -3,7 +3,7 @@ import Head from "next/head";
 import { useTranslation } from "next-i18next";
 import { Hero, Footer, MainLayout, HomeGutter } from "@gouvfr-lasuite/ui-kit";
 import { login, useAuth } from "@/features/auth/Auth";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import logoGouv from "@/assets/logo-gouv.svg";
 import banner from "@/assets/home/banner.png";
 import { HeaderRight } from "@/features/layouts/components/header/Header";
@@ -18,12 +18,11 @@ import { LeftPanelMobile } from "@/features/layouts/components/left-panel/LeftPa
 import { SESSION_STORAGE_REDIRECT_AFTER_LOGIN_URL } from "@/features/api/fetchApi";
 import { useThemeCustomization } from "@/hooks/useThemeCustomization";
 import { Feedback } from "@/features/feedback/Feedback";
+
+
 export default function HomePage() {
   const { t } = useTranslation();
   const { user } = useAuth();
-  const { config } = useConfig();
-
-  const footerCustommization = useThemeCustomization("footer");
 
   // Redirect to the attempted url if it exists, otherwise redirect to the last visited item.
   useEffect(() => {
@@ -66,8 +65,53 @@ export default function HomePage() {
     return null;
   }
 
+  return <HomePageContent />;
+}
+
+/**
+ * If the FRONTEND_EXTERNAL_HOME_URL is set, we redirect to it.
+ * Otherwise, we display the home page.
+ * 
+ * Redirection to FRONTEND_EXTERNAL_HOME_URL is done in this component
+ * to avoid conflicts with the useEffect and redirection logic in the HomePage component.
+ * 
+ * HomePage: if there is a user, redirect to the explorer.
+ * HomePageContent: if the FRONTEND_EXTERNAL_HOME_URL is set, we redirect to it.
+ *                  Otherwise, we display the home page.
+ */
+const HomePageContent = () => {
+  const { t } = useTranslation();
+  const { config } = useConfig();
+  const footerCustommization = useThemeCustomization("footer");
+  const [redirectFailed, setRedirectFailed] = useState(false)
+
+  useEffect(() => {
+    const checkSiteAndRedirect = async () => {
+      if (!config?.FRONTEND_EXTERNAL_HOME_URL) {
+        return;
+      }
+      try {
+        // Make sure the site is reachable before redirecting. Resilience.
+        await fetch(config.FRONTEND_EXTERNAL_HOME_URL, {
+          method: 'HEAD', // Use HEAD to avoid downloading the full page
+          mode: 'no-cors', // Needed for cross-origin requests
+        })
+        window.location.replace(config.FRONTEND_EXTERNAL_HOME_URL)
+      } catch (error) {
+        console.warn('Site is not reachable:', error)
+        setRedirectFailed(true)
+      }
+    }
+
+    checkSiteAndRedirect()
+  }, [config?.FRONTEND_EXTERNAL_HOME_URL])
+
+  if (config?.FRONTEND_EXTERNAL_HOME_URL && !redirectFailed) {
+    return null;
+  }
+
   return (
-    <>
+    <HomePageLayout>
       <Head>
         <title>{t("app_title")}</title>
         <meta name="description" content={t("app_description")} />
@@ -104,30 +148,43 @@ export default function HomePage() {
         />
       </HomeGutter>
       <Footer {...footerCustommization} />
-    </>
+    </HomePageLayout>
   );
 }
 
+
+const HomePageLayout = ({children}: {children: React.ReactNode}) => {
+  return (
+    <MainLayout
+      enableResize
+      hideLeftPanelOnDesktop={true}
+      leftPanelContent={<LeftPanelMobile />}
+      icon={
+        <div className="drive__header__left">
+          <img src={logoGouv.src} alt="" />
+          <div className="drive__header__logo" />
+          <Feedback />
+        </div>
+      }
+      rightHeaderContent={<HeaderRight />}
+    >
+      {children}
+      <Toaster />
+    </MainLayout>
+  );
+}
+
+/**
+ * Only context stuff, containing Auth, etc ...
+ * Do not include any interface related component here as if there is
+ * an external home url defined, we do not want blinking effects happening
+ * before redirection.
+ */
 HomePage.getLayout = function getLayout(page: React.ReactElement) {
   return (
     <div className="drive__home drive__home--feedback">
       <GlobalLayout>
-        <MainLayout
-          enableResize
-          hideLeftPanelOnDesktop={true}
-          leftPanelContent={<LeftPanelMobile />}
-          icon={
-            <div className="drive__header__left">
-              <img src={logoGouv.src} alt="" />
-              <div className="drive__header__logo" />
-              <Feedback />
-            </div>
-          }
-          rightHeaderContent={<HeaderRight />}
-        >
-          {page}
-          <Toaster />
-        </MainLayout>
+        {page}
       </GlobalLayout>
     </div>
   );