+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- 🏗️ (ds_proxy) introduce how to use ds_proxy with Drive
+
 ## [v0.11.1] - 2026-01-13
 
 ### Fixed