+++ CHANGELOG.md
@@ -10,11 +10,15 @@ and this project adheres to
 
 ### Added
 
-- 🏗️ (ds_proxy) introduce how to use ds_proxy with Drive
+- 🏗️(ds_proxy) introduce how to use ds_proxy with Drive
 
 ### Changed
 
 - 🔥(backend) remove usage of atomic transaction for item creation
+
+### Security
+
+- 🔒️(backend) prevent mismatch mimetype between object storage and application
 
 ## [v0.11.1] - 2026-01-13