+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [v0.12.0] - 2026-02-06
+
 ### Added
 
 - 🏗️(ds_proxy) introduce how to use ds_proxy with Drive
@@ -265,7 +267,8 @@ and this project adheres to
 - 🌐(front) add english translation for rename modal
 - 🐛(global) fix wrong Content-Type on specific s3 implementations
 
-[unreleased]: https://github.com/suitenumerique/drive/compare/v0.11.1...main
+[unreleased]: https://github.com/suitenumerique/drive/compare/v0.12.0...main
+[v0.12.0]: https://github.com/suitenumerique/drive/releases/v0.12.0
 [v0.11.1]: https://github.com/suitenumerique/drive/releases/v0.11.1
 [v0.11.0]: https://github.com/suitenumerique/drive/releases/v0.11.0
 [v0.10.1]: https://github.com/suitenumerique/drive/releases/v0.10.1
@@ -281,3 +284,4 @@ and this project adheres to
 [v0.2.0]: https://github.com/suitenumerique/drive/releases/v0.2.0
 [v0.1.1]: https://github.com/suitenumerique/drive/releases/v0.1.1
 [v0.1.0]: https://github.com/suitenumerique/drive/releases/v0.1.0
+## [v0.11.1] - 2026-01-13